#75 - fix

+++ app/Eloquent/Helpers/ColorGenerator.php
@@ -16,7 +16,7 @@ class ColorGenerator
         return $colors[$index];
     }
 
-    protected static function calculateHash(string $text): float
+    protected static function calculateHash(string $text): int
     {
         $hash = 0;
 
@@ -25,7 +25,7 @@ class ColorGenerator
         }
 
         for ($i = 0; $i < mb_strlen($text); $i++) {
-            $hash = abs(($hash << 2) - $hash) + mb_ord($text[$i]);
+            $hash = abs((int) (($hash << 2) - $hash) + mb_ord($text[$i]));
         }
 
         return $hash;
